Consolidate interdependent files.
There’s a circular relationship between the generic color parser and the rgb and
hsl constructors: the color parser instantiates an rgb or hsl color with the
parsed channel values, and the rgb and hsl constructors (when called with a
single argument) applies the color parser to first create an RGB color.

There’s also a circular relationship between the Lab and HCL color spaces, since
the latter is a cylindrical transformation of the former.

Some module loaders (notably Babel; see #22) apparently cannot handle circular
dependencies between files. So now the files are merged.