Implicit parsing of undefined channels.
CSS has no way of expressing an unset channel, and so for example “transparent”
is defined arbitrarily as rgba(0, 0, 0, 0). This module, in contrast, uses NaN
channels to distinguish a channel that is unset. Therefore when parsing CSS
color specifiers, we must record the unset channels (such as the hue being
undefined when the color is black).