Support CSS4 commaless syntax and units
CSS4 allows omitting the commas from the rgb() and hsl() functions, and
defines rgba() and hsla() as aliases of the former two functions. It
also allows more units: percentage for the opacity and
radian/gradian/turn for the hue. All these are already implemented in
current web browsers, even though HWB (which does not support the
"legacy" CSS3 syntax) is not.